+++ docs/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+v3.0.0
+------
+
+* Added `swift.update_settings` function for updating swift module settings. 
+  Settings may no longer be changed at module level.
+
 v2.3.0
 ------